Abort on panic
Maybe we should catch panics instead of aborting, but that would
introduce an (not noticeable?) overhead and make the API a little bit
more complicated (would need to return Result-like types). I think for
now, aborting is okay, just don't cause it ...
The "C-unwind" does not allow to start unwinding in Rust and end it in,
e.g., C++. Instead it has to go back to Rust.